move doubleLineSpitter to tablehelpers.stanzaSplitter
